refactor: domain 변화 적용
